+++ cmd/ast.go
@@ -2,15 +2,11 @@ package cmd
 
 import (
 	"asana-report/model"
+	"asana-report/service"
 	"asana-report/util"
-	"fmt"
-	"net/http"
 
-	"github.com/go-resty/resty/v2"
 	"github.com/spf13/cobra"
 )
-
-var client = resty.New()
 
 var isFullReport bool
 
@@ -19,9 +15,8 @@ var astCmd = &cobra.Command{
 	Short: "Get task status",
 	Long:  ``,
 	Run: func(cmd *cobra.Command, args []string) {
+		service.InitService()
 		var response model.AstResponse
-
-		client.SetHostURL(util.GLOBAL_CONFIG.AsanaUrl)
 
 		workspaceId := util.GLOBAL_CONFIG.WorkspaceId
 		token := util.GLOBAL_CONFIG.Token
@@ -30,13 +25,13 @@ var astCmd = &cobra.Command{
 
 		cmd.Println("================================================")
 
-		workspace := getWorkspace(workspaceId, token)
+		workspace := service.GetWorkspace(workspaceId, token)
 
 		for _, project := range workspace.Project {
 			cmd.Printf("Project: %v in progress...\n", project.Name)
 
 			taskCompleted := 0
-			tasks := getTasks(project.Gid, token)
+			tasks := service.GetTasks(project.Gid, token)
 
 			for _, task := range tasks.Data {
 				if task.Completed {
@@ -63,55 +58,6 @@ var astCmd = &cobra.Command{
 func init() {
 	rootCmd.AddCommand(astCmd)
 	astCmd.Flags().BoolVarP(&isFullReport, "full-report", "f", false, "add -f tag for print full report (default is short report)")
-}
-
-func getTasks(projectId string, token string) model.Tasks {
-	tasks := model.Tasks{}
-
-	response, err := client.R().
-		EnableTrace().
-		SetAuthToken(token).
-		SetResult(&tasks).
-		SetPathParams(map[string]string{
-			"project_id": projectId,
-		}).
-		SetQueryParams(map[string]string{
-			"opt_fields": "completed,name",
-		}).
-		Get("projects/{project_id}/tasks")
-	if err != nil {
-		panic(err)
-	}
-
-	if response.StatusCode() != http.StatusOK {
-		errorString := fmt.Sprintf("Something wrong from asana status code is %v at getTasks()\n", response.StatusCode())
-		panic(errorString)
-	}
-
-	return tasks
-}
-
-func getWorkspace(workspaceId string, token string) model.Workspace {
-	workspace := model.Workspace{}
-
-	response, err := client.R().
-		EnableTrace().
-		SetAuthToken(token).
-		SetResult(&workspace).
-		SetQueryParams(map[string]string{
-			"workspace": workspaceId,
-		}).
-		Get("projects")
-	if err != nil {
-		panic(err)
-	}
-
-	if response.StatusCode() != http.StatusOK {
-		errorString := fmt.Sprintf("Something wrong from asana status code is %v at getWorkspace()\n", response.StatusCode())
-		panic(errorString)
-	}
-
-	return workspace
 }
 
 func getSumCompletedAndTask(astList []model.Ast) (int, int) {
